Testing Inception

+++ lib/preprocessing/inception_preprocessing.py
@@ -388,18 +388,18 @@ def preprocess_for_eval(image, lidar, height, width, pool_size=None,
 
     def f_height(image,lidar):
       image = tf.slice(image,
-                       [(image_height-image_width)/2,0,0],
+                       [tf.to_int32((image_height-image_width)/2),0,0],
                        [image_width,image_width,3])
       lidar = tf.slice(lidar,
-                       [(image_height-image_width)/2,0,0],
+                       [tf.to_int32((image_height-image_width)/2),0,0],
                        [image_width,image_width,1])
       return image,lidar
     def f_width(image,lidar):
       image = tf.slice(image,
-                       [0,(image_width-image_height)/2,0],
+                       [0,tf.to_int32((image_width-image_height)/2),0],
                        [image_height,image_height,3])
       lidar = tf.slice(lidar,
-                       [0,(image_width-image_height)/2,0],
+                       [0,tf.to_int32((image_width-image_height)/2),0],
                        [image_height,image_height,1])
       return image,lidar
